+++ docs/release-notes.md
@@ -1,17 +1,21 @@
 # msaDocModels Release Notes
 ## Possible future features:
 
+# 0.0.101
+
+- added supporting pydantic v2
+
 # 0.0.100
 
-- add Input and DTO models for Spellcheck 
+- add Input and DTO models for Spellcheck
 
 # 0.0.98
 
-- fix algorithm to DocClassifier Input 
+- fix algorithm to DocClassifier Input
 
 # 0.0.97
 
-- add algorithm to DocClassifier Input 
+- add algorithm to DocClassifier Input
 - change dto model for DocClassifier
 
 # 0.0.96